+++ packages/cater-build/client/cater/head-link.js
@@ -1,0 +1,6 @@
+// Copyright Jon Williams 2017-2018. See LICENSE file.
+import 'react';
+
+// Global styles are rendered on the server-side, so we have a stubbed
+// out component here.
+export default () => false;